Use RequestBufferStore for request buffers
Rename to make room for editor buffers.
